Fix timezone issue on upcoming events
